monitoring/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.3
django-filter==22.1
django-geojson==4.0.0
django-leaflet==0.28.3
djangorestframework==3.14.0
GDAL @ file:///C:/Users/Clevo/Downloads/django-leaflet-master/GDAL-3.4.3-cp38-cp38-win_amd64.whl
importlib-metadata==5.1.0
jsonfield==3.1.0
Markdown==3.4.1
Pillow==9.3.0
psycopg2-binary==2.9.5
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6
zipp==3.11.0
